# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Nutrition tracking API.\n\nError code catalog:\n- See docs/openapi-error-codes.md in repository root.",
        "title": "Nutrition API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/foods": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "List foods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by food name (case-insensitive, partial match)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.FoodResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Create food",
                "parameters": [
                    {
                        "description": "Food payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFoodRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.FoodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/foods/by-barcode/{barcode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Get food by barcode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Food barcode",
                        "name": "barcode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.FoodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/foods/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Get food by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.FoodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Update food",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Food update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFoodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.FoodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateFoodRequest": {
            "type": "object",
            "properties": {
                "barcode": {
                    "description": "Optional product barcode (EAN/UPC digits).",
                    "type": "string",
                    "example": "5901234123457"
                },
                "carbs_per_100g": {
                    "description": "Carbohydrate grams per 100g.",
                    "type": "number",
                    "example": 28
                },
                "fat_per_100g": {
                    "description": "Fat grams per 100g.",
                    "type": "number",
                    "example": 0.3
                },
                "kcal_per_100g": {
                    "description": "Energy in kcal per 100g.",
                    "type": "number",
                    "example": 130
                },
                "name": {
                    "description": "Human-readable food name.",
                    "type": "string",
                    "example": "Rice"
                },
                "protein_per_100g": {
                    "description": "Protein grams per 100g.",
                    "type": "number",
                    "example": 2.7
                }
            }
        },
...
        "dto.UpdateFoodRequest": {
            "type": "object",
            "properties": {
                "barcode": {
                    "description": "Optional product barcode (EAN/UPC digits).",
                    "type": "string",
                    "example": "5901234123457"
                },
                "carbs_per_100g": {
                    "description": "Optional carbohydrate grams per 100g.",
                    "type": "number",
                    "example": 28
                },
                "fat_per_100g": {
                    "description": "Optional fat grams per 100g.",
                    "type": "number",
                    "example": 0.3
                },
                "kcal_per_100g": {
                    "description": "Optional energy in kcal per 100g.",
                    "type": "number",
                    "example": 130
                },
                "name": {
                    "description": "Optional food name.",
                    "type": "string",
                    "example": "Cooked Rice"
                },
                "protein_per_100g": {
                    "description": "Optional protein grams per 100g.",
                    "type": "number",
                    "example": 2.7
                }
            }
        },
...
        "handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Stable machine-readable error code.",
                    "type": "string",
                    "example": "invalid_food_payload"
                },
                "message": {
                    "description": "Human-readable error message.",
                    "type": "string",
                    "example": "invalid food payload"
                },
                "request_id": {
                    "description": "Request ID for log correlation.",
                    "type": "string",
                    "example": "f328f2f67798/SrQv8xTxwA-000001"
                }
            }
        },
...
        "handlers.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.APIError"
                }
            }
        },
        "handlers.FoodResponse": {
            "type": "object",
            "properties": {
                "barcode": {
                    "description": "Optional product barcode.",
                    "type": "string",
                    "example": "5901234123457"
                },
                "carbs_per_100g": {
                    "description": "Carbohydrate grams per 100g.",
                    "type": "number",
                    "example": 28
                },
                "created_at": {
                    "description": "Creation timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "fat_per_100g": {
                    "description": "Fat grams per 100g.",
                    "type": "number",
                    "example": 0.3
                },
                "id": {
                    "description": "Food ID.",
                    "type": "integer",
                    "example": 1
                },
                "kcal_per_100g": {
                    "description": "Energy in kcal per 100g.",
                    "type": "number",
                    "example": 130
                },
                "name": {
                    "description": "Food name.",
                    "type": "string",
                    "example": "Rice"
                },
                "protein_per_100g": {
                    "description": "Protein grams per 100g.",
                    "type": "number",
                    "example": 2.7
                },
                "updated_at": {
                    "description": "Last update timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                }
            }
        },
...
    }
}